requirements.txt
Flask==3.0.3
PyPDF2==3.0.1
google-generativeai==0.8.3
gTTS==2.5.1
moviepy==1.0.3
Pillow==10.4.0
python-dotenv==1.0.1
numpy>=2.1.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
decorator==4.4.2
tqdm==4.66.4
